RF: Move validation to its own module
