+++ psychopy/validation/__init__.py
@@ -1,0 +1,3 @@
+"""
+This module contains classes for validating stimulus presentation times using a variety of methods.
+"""
